\name{GANPAdata-package}
\alias{GANPAdata-package}
\alias{GANPAdata}
\docType{package}
\title{
The GANPA Datasets Package
}
\description{
This is a dataset package for GANPA, which implements a network-based gene weighting approach to pathway analysis. This package includes data useful for GANPA, such as a functional association network, pathways, an expression dataset and multi-subunit proteins.
}
\details{
\tabular{ll}{
Package: \tab GANPAdata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-26\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhaoyuan Fang, Weidong Tian and Hongbin Ji

Maintainer: Zhaoyuan Fang \email{fangzhaoyuan@sibs.ac.cn}
}
\references{
Zhaoyuan Fang, Weidong Tian and Hongbin Ji. A Network-Based Gene Weighting Approach for Pathway Analysis. \emph{Submitted}.

Aravind Subramanian, Pablo Tamayo, Vamsi K. Mootha, Sayan Mukherjee, Benjamin L. Ebert, Michael A. Gillette, Amanda Paulovich, Scott L. Pomeroy, Todd R. Golub, Eric S. Lander, and Jill P. Mesirov. Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles. \emph{PNAS} 2005 102(43): 15545-15550.
}
\keyword{ package }
