\name{gNET}
\alias{gNET}
\docType{data}
\title{
gNET: A comprehensive gene functional association network
}
\description{
gNET was constructed from protein-protein interactions (PPI), co-annotation in GO Biological Process (BP), and co-expression in large scale gene expression microarray data. Several limits have been applied to these data sources to achieve specificity.
}
\usage{data(gNET)}
\format{A List of 16979 character vectors.}
\references{
Zhaoyuan Fang, Weidong Tian and Hongbin Ji. A Network-Based Gene Weighting Approach for Pathway Analysis. \emph{Submitted}.
}
\examples{
data(gNET)
}
\keyword{datasets}
