\name{gsets.msigdb.pnas}
\alias{gsets.msigdb.pnas}
\docType{data}
\title{
Functional Gene Sets Used in GSEA PNAS Publication
}
\description{
A list of 522 C2 functional gene sets, see reference for details. 
}
\usage{data(gsets.msigdb.pnas)}
\format{A List of 522 character vectors.}
\source{
http://www.broadinstitute.org/gsea/index.jsp
}
\references{
Aravind Subramanian, Pablo Tamayo, Vamsi K. Mootha, Sayan Mukherjee, Benjamin L. Ebert, Michael A. Gillette, Amanda Paulovich, Scott L. Pomeroy, Todd R. Golub, Eric S. Lander, and Jill P. Mesirov. Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles. \emph{PNAS} 2005 102(43): 15545-15550.
}
\examples{
data(gsets.msigdb.pnas)
}
\keyword{datasets}
