\name{msp.groups}
\alias{msp.groups}
\docType{data}
\title{
A List of Human Multi-subunit Proteins
}
\description{
A list of genes encoding 82 multi-subunit proteins were extracted from human genes. 
}
\usage{data(msp.groups)}
\format{A List of 82 character vectors.}
\details{
A multi-subunit protein is named in the format of Multi.XXX where XXX is a member gene picked arbitrarily from the subunit gene group. Note this naming rule is just for convenience. 
}
\references{
Zhaoyuan Fang, Weidong Tian and Hongbin Ji. A Network-Based Gene Weighting Approach for Pathway Analysis. \emph{Submitted}.
}
\examples{
data(msp.groups)
}
\keyword{datasets}
